Rework module to avoid migrations' ordering problems
Add a new configuration key (with a default value) for the Duct
module, to specify the name of the table storing applied Ragtime
migrations. This is so we can store those migrations separately from
the main application migrations, and don't interfere with them.

Otherwise this causes lots of problems because of migrations ordering,
as we merge the Quartz/Twarc migrations with the already existing ones
from the application. Because of the way Duct modules work, the
migration added by the module is always positioned in the last
position of the migrations vector. So each time we add a new
application migration, the order of the migrations changes.

The Ragtime documentation says that the migrations have to be ordered,
and any change in the order produces conflicts. Depending on the
configured migrations strategy, the conflicts will prevent the app to
start. That meant that once we add this module, we couldn't add more
migrations without producing a conflict. Using a separate table for
the module migrations avoids this problem.

Alternatively the Duct module can be ignored and the migrations
explicitly added to the list of the application migrations, as the
library migrations are now accessible in a resource directory
qualified with the library namespace. That makes it very unlikely that
they can collide with any other application migration names or paths.

Closes #2