Refactor PostgresSQL connection config to take a JDBC URL
The underlying Quartz library can use such a URL if it includes the
username and password, without needing to pass them as separate
options like we did before. At the end of the day we were taking the
individual components separately to build a JDBC URL. So better take
the JDBC URL directly and be done with it.

In addition to that, this refactoring brings the JDBC datasource
configuration inline with the rest of our JDBC-related libraries.

[Fix #3]